refactor: delete vite configiration, add next configurarion, in this case move some files
