+++ package.json
@@ -4,10 +4,10 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc -b && vite build",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
     "lint": "eslint .",
-    "preview": "vite preview",
     "format:fix": "prettier --write .",
     "prepare": "husky install",
     "predeploy": "npm run build",
@@ -19,11 +19,11 @@
     "axios": "^1.7.9",
     "gh-pages": "^6.3.0",
     "msw": "^2.7.0",
+    "next": "^15.2.0",
     "node-fetch": "^3.3.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-redux": "^9.2.0",
-    "react-router-dom": "^7.1.5",
     "util": "^0.12.5"
   },
   "devDependencies": {
@@ -36,7 +36,6 @@
     "@types/react": "^18.3.18",
     "@types/react-dom": "^18.3.5",
     "@types/testing-library__user-event": "^4.2.0",
-    "@vitejs/plugin-react": "^4.3.4",
     "axios-mock-adapter": "^2.1.0",
     "eslint": "^9.17.0",
     "eslint-config-prettier": "^10.0.1",
@@ -55,7 +54,6 @@
     "ts-jest": "^29.2.5",
     "typescript": "~5.6.2",
     "typescript-eslint": "^8.18.2",
-    "vite": "^6.0.5",
     "whatwg-fetch": "^3.6.20"
   },
   "lint-staged": {
