+++ pages/class-components/404.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import NotFoundPage from '../../components/NotFoundPage';
+
+const NotFound = () => {
+  return <NotFoundPage />;
+};
+
+export default NotFound;
