+++ pages/class-components/[id].tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { useRouter } from 'next/router';
+import Card from '../../components/Card';
+
+const CardPage = () => {
+  const router = useRouter();
+  const { id } = router.query;
+
+  return <Card id={id as string} />;
+};
+
+export default CardPage;
