+++ pages/class-components/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import Main from '../../components/Main';
+
+const ClassComponents = () => {
+  return <Main searchTerm="" />;
+};
+
+export default ClassComponents;
