--- src/main.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-import './index.css';
-import { Provider } from 'react-redux';
-import { store } from './store';
-import App from './App';
-
-createRoot(document.getElementById('root') as HTMLElement).render(
-  <Provider store={store}>
-    <App />
-  </Provider>
-);
